tests/wpt/css-tests/css-flexbox-1_dev/html/percentage-heights-000.htm

<!DOCTYPE html>
<html><head><title>CSS Flexbox: Basic percentage heights</title>
<link href="https://www.google.com/" rel="author" title="Google">
<link href="https://drafts.csswg.org/css-flexbox-1/#main-sizing" rel="help">
<meta content="dom" name="flags">
<meta content="This test checks that percentage heights on a flex item correctly resolve against the container." name="assert">
<style>
.flexbox {
    display: flex;
    background-color: #aaa;
    position: relative;
}
.flexbox :nth-child(1) {
    background-color: blue;
}
.flexbox :nth-child(2) {
    background-color: green;
}
.flexbox :nth-child(3) {
    background-color: red;
}

.flexbox > div {
    width: 40%;
    height: 40%;
}
.column {
    flex-flow: column wrap;
    width: 100px;
    height: 100px;
    align-content: flex-start;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/check-layout-th.js"></script>
</head><body onload="checkLayout('.flexbox')">
<div id="log"></div>

<div class="flexbox column">
    <div data-expected-height="40" data-offset-x="0" data-offset-y="0"></div>
    <div data-expected-height="40" data-offset-x="0" data-offset-y="40"></div>
    <div data-expected-height="40" data-offset-x="40" data-offset-y="0"></div>
</div>

<div class="flexbox column">
    <div style="margin-bottom: 10%" data-expected-height="40" data-offset-x="0" data-offset-y="0"></div>
    <div style="margin-bottom: 20%" data-expected-height="40" data-offset-x="40" data-offset-y="0"></div>
    <div data-expected-height="40" data-offset-x="40" data-offset-y="60"></div>
</div>

<div class="flexbox column">
    <div style="flex: 1; min-height: 0; max-height: 20%" data-expected-height="20" data-offset-x="0" data-offset-y="0"></div>
    <div data-expected-height="40" data-offset-x="0" data-offset-y="20"></div>
    <div data-expected-height="40" data-offset-x="0" data-offset-y="60"></div>
</div>

<div style="writing-mode: vertical-rl" class="flexbox column">
    <div data-expected-width="40" data-offset-x="60" data-offset-y="0"></div>
    <div data-expected-width="40" data-offset-x="20" data-offset-y="0"></div>
    <div data-expected-width="40" data-offset-x="60" data-offset-y="40"></div>
</div>

<div style="writing-mode: vertical-rl" class="flexbox column">
    <div style="margin-bottom: 10%" data-expected-width="40" data-offset-x="60" data-offset-y="0"></div>
    <div style="margin-bottom: 20%" data-expected-width="40" data-offset-x="20" data-offset-y="0"></div>
    <div data-expected-width="40" data-offset-x="60" data-offset-y="60"></div>
</div>

<div style="writing-mode: vertical-rl" class="flexbox column">
    <div style="flex: 1; min-width: 0; max-width: 20%" data-expected-width="20" data-offset-x="80" data-offset-y="0"></div>
    <div data-expected-width="40" data-offset-x="40" data-offset-y="0"></div>
    <div data-expected-width="40" data-offset-x="0" data-offset-y="0"></div>
</div>



</body></html>
